+++ vxutil-vertigram/src/main/kotlin/ski/gagar/vxutil/vertigram/types/InlineQueryResultVenue.kt
@@ -1,8 +1,5 @@
 package ski.gagar.vxutil.vertigram.types
 
-/**
- * Telegram type InlineQueryResultVenue
- */
 data class InlineQueryResultVenue(
     val id: String,
     val latitude: Double,
@@ -16,8 +13,8 @@ data class InlineQueryResultVenue(
     val replyMarkup: InlineKeyboardMarkup? = null,
     val inputMessageContent: InputMessageContent? = null,
     val thumbUrl: String? = null,
-    val thumbWidth: Long? = null,
-    val thumbHeight: Long? = null
-) : InlineQueryResult() {
+    val thumbWidth: Int? = null,
+    val thumbHeight: Int? = null
+) : InlineQueryResult {
     override val type: InlineQueryResultType = InlineQueryResultType.VENUE
 }
