+++ vxutil-vertigram/src/main/kotlin/ski/gagar/vxutil/vertigram/types/Invoice.kt
@@ -1,12 +1,9 @@
 package ski.gagar.vxutil.vertigram.types
 
-/**
- * Telegram type Invoice.
- */
 data class Invoice(
     val title: String,
     val description: String,
     val startParameter: String,
     val currency: String,
-    val totalAmount: Long
+    val totalAmount: Int
 )
