+++ vxutil-vertigram/src/main/kotlin/ski/gagar/vxutil/vertigram/types/LabeledPrice.kt
@@ -1,9 +1,6 @@
 package ski.gagar.vxutil.vertigram.types
 
-/**
- * Telegram type LabeledPrice.
- */
 data class LabeledPrice(
     val label: String,
-    val amount: Long
+    val amount: Int
 )
