+++ vxutil-vertigram/src/main/kotlin/ski/gagar/vxutil/vertigram/types/PhotoSize.kt
@@ -1,12 +1,9 @@
 package ski.gagar.vxutil.vertigram.types
 
-/**
- * Telegram type PhotoSize.
- */
 data class PhotoSize(
     val fileId: String,
     val fileUniqueId: String,
-    val width: Long,
-    val height: Long,
+    val width: Int,
+    val height: Int,
     val fileSize: Long? = null
 )
