+++ vxutil-vertigram/src/main/kotlin/ski/gagar/vxutil/vertigram/types/PreCheckoutQuery.kt
@@ -1,13 +1,10 @@
 package ski.gagar.vxutil.vertigram.types
 
-/**
- * Telegram type PreCheckoutQuery.
- */
 data class PreCheckoutQuery(
     val id: String,
     val from: User,
     val currency: String,
-    val totalAmount: Long,
+    val totalAmount: Int,
     val invoicePayload: String,
     val shippingOptionId: String? = null,
     val orderInfo: OrderInfo? = null
