+++ vxutil-vertigram/src/main/kotlin/ski/gagar/vxutil/vertigram/types/ProximityAlertTriggered.kt
@@ -1,10 +1,7 @@
 package ski.gagar.vxutil.vertigram.types
 
-/**
- * Telegram type ProximityAlertTriggered
- */
 data class ProximityAlertTriggered(
     val traveler: User,
     val watcher: User,
-    val distance: Long
+    val distance: Int
 )
